+++ server/data/disruption/writeup/station-closure-disruption-writeup-author.ts
@@ -1,4 +1,4 @@
-import { stations } from "@/server/data/static/stations";
+import { stations } from "@/server/entry-point/stations";
 import { StationClosureDisruptionData } from "@/server/data/disruption/data/station-closure-disruption-data";
 import { Disruption } from "@/server/data/disruption/disruption";
 import { DisruptionWriteup } from "@/server/data/disruption/writeup/disruption-writeup";
